requirements.txt
cvzone
ultralytics
hydra-core
matplotlib
numpy
opencv-python
Pillow
PyYAML
requests
scipy
torch
torchvision
tqdm
filterpy
scikit-image
lap